Agrego abm de perfiles y se usa datatables

+++ Models/Family.cs
@@ -6,7 +6,7 @@ namespace doCloud.Models
     public class Family
     {
         public string descripcion { get; set; }
-        public int familia_id { get; set; }
+        public double familia_id { get; set; }
         public List<Patent> patentes { get; set; }
         public List<DocumentType> documentos_tipos { get; set; }
